Refactor tri-state to accept promises
Rather than make requests in the component which causes all kinds of
problems when it comes to tying into Ember's provided tools for causing
model reloads, this approach accepts promises rather than functions
which generate promises. Its the same basic concept but a bit simplified
and removes the dependency on concurrency.